requirements.txt
watchdog>=3.0.0
google-generativeai>=0.3.0
click>=8.0.0
pyyaml>=6.0
rich>=13.0.0
textual>=0.50.0
